Move: HABIT_VALIDATION 이동 #66
